Module Sihl_core__.Container

Lifecycle

Every service has a lifecycle, meaning it can be started and stopped. *

module Lifecycle : sig ... end

Service

A service has a start and stop functions and a lifecycle. *

module Service : sig ... end

val start_services : Service.t list -> Lifecycle.t list Lwt.t

start_services lifecycles starts a list of service lifecycles. The order does not matter as the services are started in the order of their dependencies. (No service is started before its dependency)

val stop_services : Service.t list -> unit Lwt.t

stop_services ctx services stops a list of service lifecycles with a context ctx. The order does not matter as the services are stopped in the order of their dependencies. (No service is stopped after its dependency)

module Map : sig ... end

val collect_all_lifecycles : Lifecycle.t list -> Lifecycle.t Map.t

val top_sort_lifecycles : Lifecycle.t list -> Lifecycle.t list

val unpack : string -> ?⁠default:'a -> 'a option Stdlib.ref -> 'a